refactor: extract PolicyCommandHandler from Program.cs
Move all policy CLI subcommands (export, import, validate, diff) into
a dedicated PolicyCommandHandler static class. Program.cs delegates
to PolicyCommandHandler.Handle(options) via a single call.

- Extracts 261 lines from the 2,268-line top-level-statements file
- Program.cs reduced to 2,307 lines (from 2,568)
- New PolicyCommandHandler.cs: 275 lines, single-responsibility class
- Zero behavioral changes — all policy commands work identically
- Makes the policy subsystem independently navigable and testable
- Follows the same extraction pattern as ConsoleFormatter
